Make dark mode scrollbar overrides an optional function rather than required
