+++ packages/material-ui-system/src/index.js
@@ -25,3 +25,4 @@ export {
   extendSxProp as unstable_extendSxProp,
 } from './styleFunctionSx';
 export { default as unstable_getThemeValue } from './getThemeValue';
+export { default as darkScrollbar } from './scrollbar';
